refactor: 주최자 접수 현황 로직 분리
