+++ backend/src/main/java/com/rungo/api/domain/registration/controller/RegistrationOrganizerQueryController.java
@@ -1,11 +1,9 @@
 package com.rungo.api.domain.registration.controller;
 
-import com.rungo.api.domain.registration.dto.MyRegistrationRes;
 import com.rungo.api.domain.registration.dto.RegistrationOverviewRes;
 import com.rungo.api.domain.registration.dto.RegistrationParticipantDetailRes;
 import com.rungo.api.domain.registration.dto.RegistrationParticipantListRes;
-import com.rungo.api.domain.registration.enumtype.RegistrationStatus;
-import com.rungo.api.domain.registration.service.RegistrationReadService;
+import com.rungo.api.domain.registration.service.RegistrationOrganizerQueryService;
 import com.rungo.api.global.response.ApiResponse;
 import com.rungo.api.global.security.SecurityUser;
 import jakarta.validation.constraints.Max;
@@ -25,26 +23,9 @@ import org.springframework.web.bind.annotation.RestController;
 @Validated
 @RestController
 @RequiredArgsConstructor
-public class RegistrationsReadController {
+public class RegistrationOrganizerQueryController {
 
-    private final RegistrationReadService registrationReadService;
-
-    // 인증된 사용자의 접수 목록 조회
-    @GetMapping("/api/v1/registrations/me")
-    public ResponseEntity<ApiResponse<MyRegistrationRes>> getMyRegistrations(
-            @AuthenticationPrincipal SecurityUser user,
-            @RequestParam(required = false) RegistrationStatus status,
-            @RequestParam(defaultValue = "0") @Min(value = 0, message = "page는 0 이상이어야 합니다.") int page,
-            @RequestParam(defaultValue = "20") @Min(value = 1, message = "size는 1 이상이어야 합니다.")
-            @Max(value = 100, message = "size는 100 이하여야 합니다.") int size
-    ) {
-        // 신청일 기준 최신순 정렬
-        Pageable pageable = PageRequest.of(page, size, Sort.by(Sort.Order.desc("appliedAt"), Sort.Order.desc("id")));
-
-        MyRegistrationRes result = registrationReadService.getMyRegistrations(user.getId(), status, pageable);
-
-        return ResponseEntity.ok(ApiResponse.ok(result));
-    }
+    private final RegistrationOrganizerQueryService registrationOrganizerQueryService;
 
     // 주최자 - 접수 요약 조회
     @GetMapping("/api/v1/organizer/marathons/{id}/registrations/summary")
@@ -52,7 +33,7 @@ public class RegistrationsReadController {
             @AuthenticationPrincipal SecurityUser organizer,
             @PathVariable("id") Long marathonId
     ) {
-        RegistrationOverviewRes result = registrationReadService.getRegistrationOverview(
+        RegistrationOverviewRes result = registrationOrganizerQueryService.getRegistrationOverview(
                 organizer.getId(), marathonId
         );
 
@@ -73,7 +54,7 @@ public class RegistrationsReadController {
         // 신청일 최신순, 같은 신청일이면 id 역순으로 고정
         Pageable pageable = PageRequest.of(page, size, Sort.by(Sort.Order.desc("appliedAt"), Sort.Order.desc("id")));
 
-        RegistrationParticipantListRes result = registrationReadService.getMarathonParticipants(
+        RegistrationParticipantListRes result = registrationOrganizerQueryService.getMarathonParticipants(
                 organizer.getId(), marathonId, courseId, name, pageable
         );
 
@@ -87,7 +68,7 @@ public class RegistrationsReadController {
             @PathVariable Long marathonId,
             @PathVariable Long registrationId
     ) {
-        RegistrationParticipantDetailRes result = registrationReadService.getMarathonParticipantDetail(
+        RegistrationParticipantDetailRes result = registrationOrganizerQueryService.getMarathonParticipantDetail(
                 organizer.getId(), marathonId, registrationId
         );
